requirements.txt
fastapi
uvicorn
scikit-learn
numpy
pandas
python-multipart
PyPDF2
python-docx
spacy